refactor(racingcar): 패키지 구조 변경
클래스를 domain, view 패키지로 분류
