+++ src/test/java/racingcar/CarTest.java
@@ -1,9 +1,9 @@
 package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.Car;
 
 import static org.assertj.core.api.Assertions.*;
